refactor(View): 각 Controller에서 사용하는 View를 개별로 만든다V C 의 의  각 에서 사용하는 V르로 를 개별로 만든다
