+++ src/Controller/WebController/LottoResultController.js
@@ -1,4 +1,5 @@
 import LottoList from "../../Model/LottoList.js";
+import LottoResultView from "../../View/WebView/LottoResultView.js";
 
 import { AMOUNT_PRICE } from "../../constants/lottoConstants.js";
 
@@ -7,22 +8,20 @@ import StatisticsResultController from "./StatisticsResultController.js";
 class LottoResultController {
   view;
   model;
-  constructor(view) {
-    this.view = view;
+  constructor() {
+    this.view = new LottoResultView();
     this.model = {};
   }
   run(amount) {
     const lottoList = new LottoList(amount);
     this.model.lottoList = lottoList;
 
-    this.view.outputView.renderLottoResult({ amount, lottoList });
+    this.view.render({ amount, lottoList });
     this.#inputLottoNumber();
   }
   #inputLottoNumber() {
-    this.view.inputView.readLottoNumber((winningNumbers, bonusNumber) => {
-      const statisticsResultController = new StatisticsResultController(
-        this.view,
-      );
+    this.view.readLottoNumber((winningNumbers, bonusNumber) => {
+      const statisticsResultController = new StatisticsResultController();
 
       statisticsResultController.run(
         winningNumbers,
